fixup - move GoogleFlightRecorder to gcp_common
